Split nimare/meta/cbmr.py into a package
Phase 1, part one: mechanical move, one class per module, no logic changed.
Mirrors the existing nimare/meta/cbma/ package precedent.

    _torch.py      12   the optional-dependency guard, in one place
    _helpers.py    63   constants and small shared functions
    results.py    274   CBMRResult
    estimator.py 1166   CBMREstimator
    inference.py 2268   CBMRInference
    __init__.py    32   public surface, plus the private names importers reach for

inference.py is still large; Phase 4 is what actually shrinks it, by lifting the
~490 lines of sandwich covariance out into strategy objects rather than methods.

Import order is _helpers <- results <- estimator <- inference. The one cycle --
CBMRResult.get_inference constructs a CBMRInference -- is broken by importing it
inside that method, which is noted at the import.

Three test changes were unavoidable, all mechanical:

- Two monkeypatch targets addressed nimare.meta.cbmr as the namespace where a
  name is looked up. It now only re-exports them, so the patches move to the
  defining modules. Monkeypatching is white-box; the assertions are unchanged.
- test_meta_package_defers_cbmr_import cleared sys.modules by exact name, which
  worked only while cbmr was a leaf module. Popping the parent while leaving its
  submodules let a later import bind a fresh parent to stale children, so
  dotted-path patches in subsequent tests addressed different module objects
  than the classes under test. It now clears by prefix and restores the original
  objects afterwards, which it never did before.

164 tests pass, including test_studyset_execution.py, which imports CBMREstimator
through the package path.
